move delete pattern to redis file
